+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * **Breaking**: dropped compatibility for Nix versions below 2.18.1
 * **Breaking**: dropped compatibility for nixpkgs-23.05.
 
+### Fixed
+* Workspace inheritance of `lints` in git dependencies is now correctly handled
+
 ## [0.15.1] - 2023-11-30
 
 ### Changed